Got rid of FullSampleView
